Add XAudio2 sample 'HelloXAudio2' (#341)
* Update AudioBuffer.cs

Fixed a bug in the static Create method where memory allocation was being provided data length when it requires size in bytes. Because the data is generic, sizeInBytes is the correct value to provide, and it is already calculated in the method.

* Add XAudio2 sample 'HelloXAudio2'

A minimalist sample demonstrating how to initialize XAudio2, load a wave file into an AudioBuffer and play it.